refactor(wish): 위시 api 모듈 분리
